(WIP) rebalance and distribute, begin refactor
