+++ storage/validator/__init__.py
@@ -24,3 +24,10 @@ from . import encryption
 from . import database
 from . import reward
 from . import bonding
+from . import network
+from . import rebalance
+from . import challenge
+from . import retrieve
+from . import distribute
+from . import store
+from . import forward
